@@ -15,35 +15,33 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.5+9
+version: 1.1.0+10
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^1.12.0
-  intl: ^0.17.0
+  firebase_core: ^2.15.0
+  intl: ^0.18.0
   uuid: ^3.0.4
   shared_preferences: ^2.0.6
-  charts_flutter: ^0.11.0
   csv: ^5.0.0
-  share_plus: ^3.0.1
-  in_app_purchase: ^1.0.6
-  in_app_purchase_ios: ^0.1.2+2
+  share_plus: ^7.2.0
+  in_app_purchase: ^3.1.7
   url_launcher: ^6.0.9
-  flutter_secure_storage: ^7.0.1
+  flutter_secure_storage: ^8.0.0
   in_app_review: ^2.0.2
 
   # Add the dependency for the FlutterFire plugin for Google Analytics
-  firebase_analytics: ^9.1.1
+  firebase_analytics: ^10.4.4
 
   # Add the dependencies for any other Firebase products you want to use in your app
   # For example, to use Firebase Authentication and Cloud Firestore
   # firebase_auth: ^0.14.0+5
   # firebase_auth_ui: ^1.0.1
-  cloud_firestore: ^3.1.9
+  cloud_firestore: ^4.8.3
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -51,11 +49,12 @@ dependencies:
   cupertino_icons: ^1.0.0
   path_provider: ^2.0.2
   sqflite: ^2.0.2
+  syncfusion_flutter_charts: ^22.1.34
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: "^0.9.0"
+  flutter_launcher_icons: ^0.13.1
 
 flutter_icons:
   android: true
