@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
 class CustomColor {
-  static final correctGreen = Colors.green[900];
-  static final incorrectRed = Colors.red[700];
-  static final primaryColor = Colors.blue[900];
-  static final backgroundColor = Colors.lightBlue[200];
-  static final disabledButton = Colors.grey[600];
-  static final selectedButton = Colors.purple[400];
+  static final correctGreen = Colors.green[900]!;
+  static final incorrectRed = Colors.red[700]!;
+  static final primaryColor = Colors.blue[900]!;
+  static final backgroundColor = Colors.lightBlue[200]!;
+  static final disabledButton = Colors.grey[600]!;
+  static final selectedButton = Colors.purple[400]!;
 }
